Simple Chat-Bot using neural network
This is a simple chat-bot which can get some input from user then convert it to a bag of words to get a prediction from the model. Find the most probable class and pick a response from the class. To take out the data we want from our json file, we need all of the nique words in our patterns.
By looping through the json data and extract the data we want. For each pattern we will turn it into a list of words using nltk.word_tokenizer rather than having them as strings. Then add each pattern into our docs_x list and it's associated tag into the docs_y list.
To develop the model I used a fairly standard feed-forward neural network with two hidden layers. The goal of my network is to look at a bag of words and give a class that it belongs.